+++ src/EyesService.js
@@ -16,57 +16,62 @@ class EyesService {
    * @param {Configuration} [config]
    */
   constructor(config) {
-    this.eyes = new Eyes();
+    this._eyes = new Eyes();
+
+    this._appName = null;
   }
 
 
   beforeSession(config, caps) {
     const eyesConfig = config.eyes;
     if (eyesConfig) {
-      this.eyes.setConfiguration(eyesConfig);
+      this._eyes.setConfiguration(eyesConfig);
+      this._appName = this._eyes.getConfiguration().getAppName();
 
       if (!process.env.APPLITOOLS_API_KEY) {
         process.env.APPLITOOLS_API_KEY = eyesConfig.apiKey;
       }
     }
-    this.eyes.setHideScrollbars(true);
+    this._eyes.setHideScrollbars(true);
   }
 
 
   before(caps) {
     browser.addCommand('eyesCheck', (title, checkSettings = Target.window().fully()) => {
-      return this.eyes.check(title, checkSettings);
+      return this._eyes.check(title, checkSettings);
     });
 
     browser.addCommand('eyesCheckWindow', (title, checkSettings) => {
-      return this.eyes.check(title, checkSettings);
+      return this._eyes.check(title, checkSettings);
     });
 
     browser.addCommand('eyesGetConfiguration', () => {
-      return this.eyes.getConfiguration();
+      return this._eyes.getConfiguration();
     });
 
   }
 
 
   async beforeTest(test) {
-    const appName = this.eyes.getConfiguration().getAppName() || test.parent;
-    const testName = test.title;
-
-    let viewport = DEFAULT_VIEWPORT;
-    if (this.eyes.getConfiguration().getViewportSize()) {
-      viewport = this.eyes.getConfiguration().getViewportSize();
+    if (!this._appName) {
+      this._eyes.getConfiguration().setAppName(test.parent);
     }
 
-    await global.browser.call(() => this.eyes.open(global.browser, appName, testName, viewport));
+    this._eyes.getConfiguration().setTestName(test.title);
+
+    if (!this._eyes.getConfiguration().getViewportSize()) {
+      this._eyes.getConfiguration().setViewportSize(DEFAULT_VIEWPORT);
+    }
+
+    await global.browser.call(() => this._eyes.open(global.browser));
   }
 
 
   async afterTest(exitCode, config, capabilities) {
     try {
-      const result = await browser.call(() => this.eyes.close(false));
+      const result = await browser.call(() => this._eyes.close(false));
     } catch (e) {
-      await browser.call(() => this.eyes.abortIfNotClosed());
+      await browser.call(() => this._eyes.abortIfNotClosed());
     }
   }