+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [unversioned] 
 ### Fixed
 - Setting viewportSize through configuration. [Ticket_1168](https://trello.com/c/yPqI3erm)
+- Set appName from describe tag. [Ticket_1174](https://trello.com/c/gIlKtwZU)
+
 ### Added
 - This changelog file.